Combine Results from Different Folds

Applies Combiners: functions that collapse across a list of similarly structured results, to a list of such lists.

combine_results(results, combiners = NULL, smart_combiners = TRUE)

Arguments

results

(list) - a list of lists, corresponding to each result, with the inner lists corresponding to results from each fold.

combiners

(list) - a list with the same names results, containing combiner function names or functions for each result.

smart_combiners

(logical) - if combiners are missing, should they be guessed from the data type of the results.

Value

A list of combined results.

Details

In theory you should never call this function directly, because it is called automatically by cross_validate. The defaults, combiners guessed based on data type, should work in most cases.

See also

Combiners