Fold Helpers

Accessors and indexers for the different parts of a fold.

training(x = NULL, fold = NULL)

validation(x = NULL, fold = NULL)

fold_index(x = NULL, fold = NULL)

Arguments

x
an object to be indexed by a training set, validation set, or fold index. If missing, the index itself will be returned.

fold
Fold; the fold used to do the indexing. If missing, fold will be pulled from the calling environment, if available.

Value

The elements of x corresponding to the indexes, or the indexes themselves if x is missing.

See also

make_fold