origami_SuperLearner

SuperLearner implemented using orgami cross-validation. Leverages a lot of code from Eric Polley's SuperLearner package. Because of it's based on origami, we get two features for free: foreach based parallelization, and support for arbitrary cross-validation schemes.

cv_SL(fold, Y, X, SL.library, family, obsWeights, id, ...)

origami_SuperLearner(Y, X, newX = NULL, SL.library, family = gaussian(),
  obsWeights = rep(1, length(Y)), id = NULL, folds = NULL,
  method = method.NNLS(), cvfun = cv_SL, control = list(), ...)

Arguments

fold

a Fold to be passed to cv_SL.

Y

vector of outcomes.

X

vector of covariates.

SL.library

character vector of learner names.

family

Either gaussian() or binomial() depending on if Y is binary or continuous.

obsWeights

vector of weights.

id

vector of ids.

...

other arguments passed to the underlying call to cross_validate

newX

currently ignored.

folds

a list of Folds. See make_folds. If missing, will be created based on Y and cluster_ids.

method

a combination method. Typically either method.NNLS or method.NNloglik.

cvfun

the function to be run at each cross-validation step. Changing this allows, for example, dynamic creation of fold-specific data. Must have same prototype as cv_SL

See also

predict.origami_SuperLearner

Examples

library(origami)
library(SuperLearner)

#> Loading required package: nnls

#> Super Learner

#> Version: 2.0-22

#> Package created on 2016-12-10

N <- 200
p <- 6
X <- matrix(rnorm(N * p), N, p)
X <- as.data.frame(X)
Y <- rbinom(N, 1, plogis(0.2 * X[, 1] + 0.1 * X[, 2] - 0.2 * X[, 3] + 0.1 * X[, 3] * X[, 4] - 0.2 * abs(X[, 4])))

SL.library <- c("SL.glmnet", "SL.glm", "SL.knn_stored", "SL.gam", "SL.mean")

folds <- make_folds(X)

# Original SuperLearner convert folds into the form SuperLearner will accept
vfolds <- lapply(folds, function(fold) {
    validation(fold = fold)
})

set.seed(1)  #cv in glmnet is randomized so we have to set a seed to consistent results
sl_result <- SuperLearner(Y = Y, X = X, SL.library = SL.library, method = method.NNLS(), family = binomial(), cvControl = list(validRows = vfolds))

#> Loading required package: gam

#> Loading required package: splines

#> Loading required package: foreach

#> foreach: simple, scalable parallel programming from Revolution Analytics
#> Use Revolution R for scalability, fault tolerance and more.
#> http://www.revolutionanalytics.com

#> Loaded gam 1.14-4

#> Loading required package: glmnet

#> Loading required package: Matrix

#> Loaded glmnet 2.0-10

#> Loading required package: class

# Origami SuperLearner
set.seed(1)  #cv in glmnet is randomized so we have to set a seed to consistent results
osl_result <- origami_SuperLearner(Y = Y, X = X, SL.library = SL.library, method = method.NNLS(), family = binomial(), folds = folds)

# compare results
equal_values <- function(e1, e2) {
    all.equal(e1, e2, check.attributes = FALSE, use.names = FALSE)
}
equal_values(sl_result$cvRisk, osl_result$cvRisk)

#> [1] TRUE

equal_values(sl_result$coef, osl_result$coef)

#> [1] TRUE

equal_values(sl_result$SL.predict, predict(osl_result$fullFit, newdata = X)$pred)

#> [1] TRUE

####################### CV.SuperLearner
set.seed(1)
cv_sl_result <- CV.SuperLearner(Y = Y, X = X, SL.library = SL.library, method = "method.NNLS", family = binomial())

# NB: not yet feature complete with CV.SuperLearner
set.seed(1)
ocv_sl_result <- origami_CV.SuperLearner(Y = Y, X = X, SL.library = SL.library, method = method.NNLS(), family = binomial())

summary(cv_sl_result)

#> 
#> Call:  
#> CV.SuperLearner(Y = Y, X = X, family = binomial(), SL.library = SL.library,  
#>     method = "method.NNLS") 
#> 
#> Risk is based on: Mean Squared Error
#> 
#> All risk estimates are based on V =  10 
#> 
#>          Algorithm     Ave        se     Min     Max
#>      Super Learner 0.25019 0.0050448 0.24073 0.26502
#>        Discrete SL 0.25250 0.0062407 0.24170 0.28658
#>      SL.glmnet_All 0.24897 0.0039209 0.24170 0.26068
#>         SL.glm_All 0.25365 0.0075989 0.22095 0.28888
#>  SL.knn_stored_All 0.26710 0.0106315 0.22350 0.32050
#>         SL.gam_All 0.25425 0.0090387 0.23228 0.28658
#>        SL.mean_All 0.24809 0.0035556 0.23985 0.25309

print(ocv_sl_result)

#>                     Ave          se       Min       Max
#> SuperLearner  0.2560788 0.005288000 0.2420443 0.2954851
#> SL.glmnet     0.2512630 0.004435932 0.2380669 0.2878943
#> SL.glm        0.2583283 0.007832351 0.2252227 0.2972540
#> SL.knn_stored 0.2710000 0.013055650 0.1770000 0.3380000
#> SL.gam        0.2575505 0.010081501 0.2120215 0.3088203
#> SL.mean       0.2504321 0.004460794 0.2368827 0.2871605

preds <- ocv_sl_result$SL.predict