Make List of Folds for cross-validation

Generates a list of folds for a variety of cross-validation schemes.

make_folds(n = NULL, fold_fun = NULL, cluster_ids = NULL,
  strata_ids = NULL, ...)

Arguments

n

- either an integer indicating the number of observations to cross-validate over, or an object from which to guess the number of observations; can also be computed from strata_ids or cluster_ids.

fold_fun

- either a character vector or a function indicating the cross-validation scheme to use.

cluster_ids

- a vector of cluster ids. Clusters are treated as a unit - that is, all observations within a cluster are placed in either the training or validation set.

strata_ids

- a vector of strata ids. Strata are balanced: insofar as possible the distribution in the sample should be the same as the distribution in the training and validation sets.

...

other arguments to be passed to fold_fun.

Value

A list of Folds.

See also

link{Fold}

Other fold_funs: fold_from_foldvec, folds_bootstrap, folds_loo, folds_montecarlo, folds_resubstitution, folds_rolling_origin, folds_rolling_window, folds_vfold, make_repeated_folds